requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
aiohttp==3.10.10
